requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-Babel==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Redis & Async
redis==5.0.1
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
